Fix breakage introduced by the test_attr infrastructure
Patchwork-id: 61962
O-Subject: [PATCH RHEL6.5 BZ984006 1/3] perf python: Fix breakage introduced by the test_attr infrastructure
Bugzilla: 984006
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 984006
https://bugzilla.redhat.com/show_bug.cgi?id=984006

upstream status
===============
commit 0c6332e9d8b8f38a9d5f0cdd66c5575e76ad8995
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Thu Dec 13 16:43:04 2012 -0300

description
===========
The test_attr infrastructure hooks on the sys_perf_event_open call,
checking if a variable is set and if so calling a function to intercept
calls and do the checking.

But both the variable and the function aren't on objects that are
linked on the python binding, breaking it:

  # perf test -v 15
  15: Try 'use perf' in python, checking link problems       :
  --- start ---
  Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
  ImportError: /home/acme/git/build/perf//python/perf.so: undefined symbol: test_attr__enabled
  ---- end ----
  Try 'use perf' in python, checking link problems: FAILED!
  #

Fix it by moving the variable to one of the linked object files and
providing a stub for the function in the python.o object, that is only
linked in the python binding.

Now 'perf test' is happy again:

  # perf test 15
  15: Try 'use perf' in python, checking link problems       : Ok
  #

Signed-off-by: Rafael Aquini <aquini@redhat.com>