hda: Check validity of CORB/RIRB WP reads
Patchwork-id: 62054
O-Subject: [RHEL 6.5 PATCH 091/556] ALSA/HDA/core: hda - Check validity of CORB/RIRB WP reads
Bugzilla: 916264
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

When the HD-audio controller is disabled (e.g. via vga switcheroo) but
the driver is still accessing it, it spews floods of "spurious
response" kernel messages.  It's because CORB/RIRB WP reads 0xff, and
the driver tries to fill up until this number.

This patch changes the CORB/RIRB WP reads to word instead of byte, and
add the check of the read value.  If it's 0xffff, the controller is
supposed to be disabled, so the further action will be skipped.

Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: cc5ede3efdf1c3777efdf92124e7572c1024a7f0)
rhbz#916264