hda: Avoid auto-mute or auto-mic of retasked jacks
Patchwork-id: 62204
O-Subject: [RHEL 6.5 PATCH 241/556] ALSA/HDA/core: hda - Avoid auto-mute or auto-mic of retasked jacks
Bugzilla: 916264
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

When a jack is retasked as a different direction (e.g. a mic jack is
used as a CLFE output), such a jack shouldn't be counted as the target
for the automatic jack switching.  Skip the automute or the autoswitch
when the current pinctl direction is different from what we suppose.

Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: 0b4df931ce3502311928bf66088cd76a2b5e604f)
rhbz#916264