hda: allow "Headphone Mic" parser flag
Patchwork-id: 62411
O-Subject: [RHEL 6.5 PATCH 448/556] ALSA/HDA/core: hda - allow "Headphone Mic" parser flag
Bugzilla: 916264
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

This allows a specific mic to get the "Headphone Mic" name, in addition
to the existing "Headset Mic" name.

Also, it allows for a special mark: if the sequence number is set
to 0xc, that's an indication to prefer it for headset mic, and if it's
set to 0xd, that's an indication to prefer it for headphone mic.

Signed-off-by: David Henningsson <david.henningsson@canonical.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: cb420b1186e0090f24137855e14e6f3bf7cb769d)
rhbz#916264