hda: Don't call vmaster hook when bus->shutdown is set
Patchwork-id: 62419
O-Subject: [RHEL 6.5 PATCH 456/556] ALSA/HDA/core: hda - Don't call vmaster hook when bus->shutdown is set
Bugzilla: 916264
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The flag bus->shutdown implies that the control elements might have
been already destroyed.  When a codec is resumed at this state and
tries to call vmaster hook (e.g. in snd_hda_gen_init()), it would
refer to a non-existing object, resulting in Oops in the end.

This patch just adds a check of the flag in the caller side for
avoiding such a crash.

Though, the best would be to clear hook->sw_kctl by the destructor of
the corresponding ctl element, but vmaster uses its own private_free,
it can't be done easily.  So let it be for a while.

Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: 594813ffa741b4bcdeff85ca09d734117ac93371)
rhbz#916264