nvme: Change the definition of nvme_user_io
Patchwork-id: 62623
O-Subject: [RHEL6.5 PATCH 068/161] NVMe: Change the definition of nvme_user_io
Bugzilla: 922973
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

From: Matthew Wilcox <matthew.r.wilcox@intel.com>

The read and write commands don't define a 'result', so there's no need
to copy it back to userspace.

Remove the ability of the ioctl to submit commands to a different
namespace; it's just asking for trouble, and the use case I have in mind
will be addressed througha  different ioctl in the future.  That removes
the need for both the block_shift and nsid arguments.

Check that the opcode is one of 'read' or 'write'.  Future opcodes may
be added in the future, but we will need a different structure definition
for them.

The nblocks field is redefined to be 0-based.  This allows the user to
request the full 65536 blocks.

Don't byteswap the reftag, apptag and appmask.  Martin Petersen tells
me these are calculated in big-endian and are transmitted to the device
in big-endian.

Signed-off-by: Matthew Wilcox <matthew.r.wilcox@intel.com>
(cherry picked from commit 6c7d49455ceb63064f992347d9185ff5bf43497a)

BZ 922973

Signed-off-by: Rafael Aquini <aquini@redhat.com>