nvme: Add the nvme thread to the wait queue before waking it up
Patchwork-id: 62686
O-Subject: [RHEL6.5 PATCH 082/161] NVMe: Add the nvme thread to the wait queue before waking it up
Bugzilla: 922973
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

From: Matthew Wilcox <matthew.r.wilcox@intel.com>

If the I/O was not completed by a single NVMe command, we add the
bio to the congestion list and wake up the kthread to resubmit it.
But the kthread calls remove_wait_queue() unconditionally, which
will oops if it's not on the wait queue.  So add the kthread to
the wait queue before waking it up.

Signed-off-by: Matthew Wilcox <matthew.r.wilcox@intel.com>
(cherry picked from commit eac623ba7a91474a688eb5d0fcd0eaa6a56dc41c)

BZ 922973

Signed-off-by: Rafael Aquini <aquini@redhat.com>