nvme: Fix calculation of number of pages in a PRP List
Patchwork-id: 62728
O-Subject: [RHEL6.5 PATCH 089/161] Fix calculation of number of pages in a PRP List
Bugzilla: 922973
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

From: Nisheeth Bhat <nisheeth.bhat@intel.com>

The existing calculation underestimated the number of pages required
as it did not take into account the pointer at the end of each page.
The replacement calculation may overestimate the number of pages required
if the last page in the PRP List is entirely full.  By using ->npages
as a counter as we fill in the pages, we ensure that we don't try to
free a page that was never allocated.

Signed-off-by: Nisheeth Bhat <nisheeth.bhat@intel.com>
Signed-off-by: Matthew Wilcox <matthew.r.wilcox@intel.com>
(cherry picked from commit 0d1bc9125890426b52ca2de6abedd32e31722e5c)

BZ 922973

Signed-off-by: Rafael Aquini <aquini@redhat.com>