e1000e: prevent multiple calls to e1000e_down() and e1000_free_irq()
Patchwork-id: 62948
O-Subject: [RHEL6.5 PATCH 87/90] e1000e: prevent multiple calls to e1000e_down() and e1000_free_irq()
Bugzilla: 923897
RH-Acked-by: John Green <jogreene@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Nikolay Aleksandrov <nikolay@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

Resolves BZ:
https://bugzilla.redhat.com/show_bug.cgi?id=923897

Upstream status:
This is a partial backport of the following from the linux tree...

commit 23606cf5d1192c2b17912cb2ef6e62f9b11de133
Author: Rafael J. Wysocki <rjw@sisk.pl>
Date:   Sun Mar 14 14:35:17 2010 +0000

    e1000e / PCI / PM: Add basic runtime PM support (rev. 4)

    Use the PCI runtime power management framework to add basic PCI
    runtime PM support to the e1000e driver.  Namely, make the driver
    suspend the device when the link is off and set it up for generating
    a wakeup event after the link has been detected again.  [This
    feature is disabled until the user space enables it with the help of
    the /sys/devices/.../power/contol device attribute.]

    Based on a patch from Matthew Garrett.

    Signed-off-by: Rafael J. Wysocki <rjw@sisk.pl>
    Signed-off-by: David S. Miller <davem@davemloft.net>

The differences between the following patch and the upstream commit are due
to RHEL6 not having pci_dev_run_wake(), which prevents this commit from
being backported in its entirety. However, the bits that prevent multiple calls
to e1000e_down() and e1000_free_irq() appear to be of value to backport.

Part of the motive for doing this partial backport is for the benefit of
the backport of upstream's a3b87a4c69619f5366b7 which follows this patch
in the patchset. Upstream's 23606cf5d1192c2b1791 is an old patch and much has
been added to RHEL6 which made it impossible to apply (even partially without
munging, regardless of its order in this patchset. So I chose to place this
patch just prior to the patch it was benefitting.