libfcoe: Check for unusable FCFs before looking for conflicting FCFs
Patchwork-id: 62974
O-Subject: [RHEL 6.5 PATCH 19/27] libfcoe: Check for unusable FCFs before looking for conflicting FCFs
Bugzilla: 829784
RH-Acked-by: Jan Vesely <jvesely@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

When there are multiple FCFs in the fabric, and one of them becomes
unavailable, the fabric name for the unavailable FCF becomes 0 along
with FIP_FL_AVAIL getting reset. In this case, FCF selection logic does
not select any FCF as it first checks for conflicting FCFs (since fabric
name is 0, it fails the condition), instead of first checking if it is
usable or not. Fix it by first checking if FCF is usable and skip that
FCF, and go to the next one in the list to check if it can be selected.

Signed-off-by: Bhanu Prakash Gollapudi <bprakash@broadcom.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>
(cherry picked from commit 1f953b0dbc2549318afcc0a70af5542dffbce34a)

BZ: 829784
BREW: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=5822433
TESTING: Tested by myself and Intel