libfcoe: Fix fcoe_sysfs VN2VN mode
Patchwork-id: 62977
O-Subject: [RHEL 6.5 PATCH 25/27] libfcoe: Fix fcoe_sysfs VN2VN mode
Bugzilla: 829784
RH-Acked-by: Jan Vesely <jvesely@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The libfc discovery layer is being initialized in the
'create' paths for both legacy libfcoe module parameters
and fcoe_sysfs control interfaces. The problem is that
for VN2VN mode the discovery layer is initialized as if
it were in 'fabric' mode and it is not re-configured when
the mode is changed to 'vn2vn'.

This patch splits out code that needs to be initialized
once and code that can, and should be, re-configured when
the mode changes. Additionally this patch makes that change
so that the discovery layer can be reconfigured to the
libfcoe implementation when in 'vn2vn' mode.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Jack Morgan <jack.morgan@intel.com>
Reviewed-by: Bhanu Prakash Gollapudi <bprakash@broadcom.com>
(cherry picked from commit 0db0e377ab5be5d507a2fca3d78215cd2e83b974)

BZ: 829784
BREW: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=5822433
TESTING: Tested by myself and Intel