megaraid_sas: megaraid_sas driver init fails in kdump kernel
Patchwork-id: 63359
O-Subject: [RHEL6.5 e-stor PATCH] megaraid_sas: megaraid_sas driver init fails in kdump kernel
Bugzilla: 833299
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

When intel_iommu=on and iommu=pt are specified for the first kernel, the kdump kernel (megaraid_sas)
fails. This is one possible solution and it's tested by the partner, the other possibilities
are still discussed in upstream (resetting the pci-e bus and devices).
The patch below is a limited driver fix, it tries a second card reset when the first fails,
so it's expected to be safe. And nothing bad should happen to us when we take this patch,
it is not a new feature.

Tomas

Upstream post - http://www.spinics.net/lists/linux-scsi/msg67288.html

Problem: When Hardware IOMMU is on, megaraid_sas driver initialization fails in kdump kernel with LSI MegaRAID controller(device id-0x73).

Actually this issue needs fix in firmware, but for firmware running in field, this driver fix is proposed to resolve the issue.
At firmware initialization time, if firmware does not come to ready state, driver will reset the adapter and retry for firmware transition to ready state unconditionally(not only executed for kdump kernel).

Signed-off-by: Sumit Saxena <sumit.saxena@xxxxxxx>
Signed-off-by: Kashyap Desai <kashyap.desai@xxxxxxx>
Signed-off-by: Rafael Aquini <aquini@redhat.com>