mpt3sas: when async scanning is enabled then while scanning, devices are removed but their transport layer entries are not removed
Patchwork-id: 63467
O-Subject: [RHEL6.5 e-stor PATCH 13/14] mpt3sas: when async scanning is enabled then while scanning, devices are removed but their transport layer entries are not removed
Bugzilla: 839470
RH-Acked-by: Jan Vesely <jvesely@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

When Async scanning mode is enabled and device scanning is in progress,
devices should not be removed. But in actuality, devices are removed but
their transport layer entries are not removed. This causes error to add
the same device to the transport layer after host reset or diagnostic
reset.

So, in this patch, modified the code in such a way that device is not removed
when Async scanning mode is enabled and device scanning is in progress.

Signed-off-by: Sreekanth Reddy <Sreekanth.Reddy@lsi.com>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
(cherry picked from commit f5edbe77d93b99f169d752d91bf31df0432ca990)

Signed-off-by: Rafael Aquini <aquini@redhat.com>