nfs: Allow nfs_updatepage to extend a write under additional circumstances
Patchwork-id: 63476
O-Subject: [RHEL6.5 PATCH] [bz983288] NFS: Allow nfs_updatepage to extend a write under additional circumstances
Bugzilla: 983288
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Fixes bug 983288.  Based on upstream commit
c7559663e42f4294ffe31fe159da6b6a66b35d61

--
Currently nfs_updatepage allows a write to be extended to cover a full
page only if we don't have a byte range lock lock on the file... but if
we have a write delegation on the file or if we have the whole file
locked for writing then we should be allowed to extend the write as
well.

Signed-off-by: Scott Mayhew <smayhew@redhat.com>
[Trond: fix up call to nfs_have_delegation()]
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
--

Signed-off-by: Scott Mayhew <smayhew@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>