nfs: fix open() O_RDONLY|O_TRUNC in NFS4.0
Patchwork-id: 63500
O-Subject: [RHEL6.5 PATCH] [BZ#982688] nfs: fix open(O_RDONLY|O_TRUNC) in NFS4.0
Bugzilla: 982688
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Nadav Shemer <nadav@tonian.com>

This patch fixes a relatively minor bug in how NFSv4 handles the
flag combination in $SUBJECT. When passed that, the current code
returns -EBADF, which doesn't make a lot of sense since we're
not passing in a file descriptor.

The problem is that nfs4_do_setattr has some special error handling
when ATTR_OPEN is set, but since we clear it here, it never is. The
fix is to leave ATTR_OPEN set, and fix the following check to ignore
it.

This patch is not yet in mainline, but Trond has merged it into his
for-next branch and I expect it'll make 3.12.

Fixes the reproducer in this bug:

    https://bugzilla.redhat.com/show_bug.cgi?id=982688

nfs4_proc_setattr removes ATTR_OPEN from sattr->ia_valid, but later
nfs4_do_setattr checks for it

Reviewed-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>