bnx2x: ignore unknown ethtool flags
Patchwork-id: 63664
O-Subject: [RHEL6.5 PATCH 188/188] bnx2x: ignore unknown ethtool flags
Bugzilla: 915517 928435
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Jan Vesely <jvesely@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Nikolay Aleksandrov <nikolay@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=928435

The update of ethtool to 3.5 in RHEL 6.4 broke the ability to disable
LRO on bnx2x. The new ethtool passes a couple of flags to ETHTOOL_SFLAGS
that bnx2x does not understand, so it refuses the whole request with
-EINVAL.

It is not necessary for bnx2x to check explicitly for known flags in its
.set_flags() ethtool operation. They should simply be ignored.
After all, it is perfectly fine if some drivers have
 .set_flags = ethtool_op_set_flags, which ignores unknown flags too.
Be lenient for the benefit of updated ethtool userspace.