Add 'eh_deadline' to limit SCSI EH runtime
Patchwork-id: 63727
O-Subject: [RHEL6.5 PATCH e-stor 7/7] scsi: Add 'eh_deadline' to limit SCSI EH runtime
Bugzilla: 923876
RH-Acked-by: Bryn M. Reeves <bmr@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Upstream status: Posted to upstream [linux-scsi] by Hannes Reinecke,
                 awaiting inclusion by James Bottomley
	(See: http://www.spinics.net/lists/linux-scsi/msg67027.html)

Tested: Tested extensively at HSBC and EMC.  Requires the ability
        to configure FC switches to suppress RSCNs when a target
        port is set offline.

This patchs adds an 'eh_deadline' attribute to the scsi
host which limits the overall runtime of the SCSI EH.
When a command is failed the start time of the EH is stored
in 'last_reset'. If the overall runtime of the SCSI EH is longer
than last_reset + eh_deadline, the EH is short-circuited and
falls through to issue a host reset only.

(Modified for RHEL6.5 -- put rename of field in struct Scsi_Host
 under #ifndef __GENKSYMS__ for KABI.)

Signed-off-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: Ewan D. Milne <emilne@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>