cifs: allocate kvec array for cifs_readdata as a separate allocation
Patchwork-id: 63758
O-Subject: [RHEL6 PATCH 18/28] [bz877010] cifs: allocate kvec array for cifs_readdata as a separate allocation
Bugzilla: 877010
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Eventually, we're going to want to append a list of pages to
cifs_readdata instead of a list of kvecs. To prepare for that, turn
the kvec array allocation into a separate one and just keep a
pointer to it in the readdata.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
(cherry picked from commit f4e49cd2dce2ccac6feae64fbb4e90f7d8baf370)

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>