ixgbe: Limit maximum jumbo frame size to 9.5K to avoid Tx hangs
Patchwork-id: 64255
O-Subject: [PATCH RHEL6.5 020/120] ixgbe/ixgbevf: Limit maximum jumbo frame size to 9.5K to avoid Tx hangs
Bugzilla: 923901
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

Backport of upstream commit:

commit c88887e09099b7f77659cb6aa1b9d670a7632f4b
Author: Alexander Duyck <alexander.h.duyck@intel.com>
Date:   Wed Aug 22 02:04:37 2012 +0000

    ixgbe/ixgbevf: Limit maximum jumbo frame size to 9.5K to avoid Tx hangs

This patch has a chance to introduce regressions, but I think it is
proper to follow upstream.  The the other option would be to set the MTU
depending on the mode used by the driver, but Intel did not want to do
this, so I do not think we want to either.