srp: use __scsi_target_unblock
Patchwork-id: 64414
O-Subject: [Patch RHEL6 45/50] IB/srp: use __scsi_target_unblock
Bugzilla: 791284 865301 950918 983117
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

We wish to be able to set the unblocked state of the scsi target
when we unblock.  In upstream, the scsi_target_unblock() entry
allows this.  In rhel6, scsi_target_unblock() forces the state
to SDEV_RUNNING.  So, use the __ version of the function that
allows us to set the state ourselves.

Signed-off-by: Doug Ledford <dledford@redhat.com>