gfs2: go back to try locks if reservation does not fit our needs
Patchwork-id: 64819
O-Subject: [RHEL6.5 PATCH] GFS2: go back to try locks if reservation does not fit our needs
Bugzilla: 989669
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

Before this patch, the GFS2 block allocator would override the
"try lock" and make it a blocking lock if it had a block reservation
going in to the allocation. However, if that resource group was
inadequate for the reservation (i.e. too few blocks to satisfy the
request) it never switched the flags back to use a try lock. From
that point on, it used blocking glocks for all resource groups.

With this patch, the try lock is overridden ONLY for the resource
group corresponding to the reservation needed. Once it determines
the resource group is inadequate, it goes back to using a try lock
for future locks (until they've all been eliminated).

rhbz#989669

This patch was tested on the afc-east cluster with the SAS calibration
workload. Before the patch, iterations of SAS calibration would get
slower and slower unless the file system was umounted and remounted
between iterations. With the patch, the performance is as good
whether or not the file system is umounted between iterations.

This patch does not apply to upstream, because the upstream GFS2 no
longer uses "try locks" when doing block allocations.

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>