qla2xxx: Don't process RSCNs for a vport on the same physical adapter
Patchwork-id: 64927
O-Subject: [RHEL 6.5 v2 e-stor PATCH 43/50] qla2xxx: Don't process RSCNs for a vport on the same physical adapter.
Bugzilla: 927382
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Upstream Status
---------------

scsi commit id bb4cf5b73b47fe78502b16bc3dc4af612aa37ad7

Description
-----------

From: Chad Dupuis <chad.dupuis@qlogic.com>

Subject: [PATCH 43/50] [SCSI] qla2xxx: Don't process RSCNs for a vport on the same physical adapter.

Currently,the driver is processes RSCNs for each new NPIV ports that is created.
Processing the RSCN includes a name server query to see what has changed at the
name server side.  The name server query is performed by the physical port and
each virtual port on the physical adapter (since the RSCN is passed to each
virtual port for processing).  As the number of virtual ports being created
increases, this causes a lot of traffic and busies the firmware.

Processing the RSCN for a virtual port we already have a priori knowledge of is
not necessary so check the 24-bit fabric ID of the RSCN entry and skip
processing it if the RSCN is for a virtual port we already know about.

Signed-off-by: Chad Dupuis <chad.dupuis@qlogic.com>
Signed-off-by: Saurav Kashyap <saurav.kashyap@qlogic.com>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>