Replace trace_event struct array with pointer array
Patchwork-id: 65075
O-Subject: [PATCH RHEL6.5 BZ992987 1/2] tracing: Replace trace_event struct array with pointer array
Bugzilla: 992987
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 992987
https://bugzilla.redhat.com/show_bug.cgi?id=992987

upstream
========
commit e4a9ea5ee7c8812a7bf0c3fb725ceeaa3d4c2fcc
Author: Steven Rostedt <srostedt@redhat.com>
Date:   Thu Jan 27 09:15:30 2011 -0500

KABI
====
The upstream version of the patch causes KABI breakage, because
it changes data of the event section used by modules.

For this backport we created the "_ftrace_events_ptrs" section
with event definitions pointers and used it for the kernel events
data only. We keep the "_ftrace_events" section with data for both
kernel and modules.

For modules we keep current behaviour - loading data from "_ftrace_events"
seciton, so we dont break KABI. Luckily modules were never seen to be the
source of the problem.

description
===========
Currently the trace_event structures are placed in the _ftrace_events
section, and at link time, the linker makes one large array of all
the trace_event structures. On boot up, this array is read (much like
the initcall sections) and the events are processed.

The problem is that there is no guarantee that gcc will place complex
structures nicely together in an array format. Two structures in the
same file may be placed awkwardly, because gcc has no clue that they
are suppose to be in an array.

A hack was used previous to force the alignment to 4, to pack the
structures together. But this caused alignment issues with other
architectures (sparc).

Instead of packing the structures into an array, the structures' addresses
are now put into the _ftrace_event section. As pointers are always the
natural alignment, gcc should always pack them tightly together
(otherwise initcall, extable, etc would also fail).

By having the pointers to the structures in the section, we can still
iterate the trace_events without causing unnecessary alignment problems
with other architectures, or depending on the current behaviour of
gcc that will likely change in the future just to tick us kernel developers
off a little more.

The _ftrace_event section is also moved into the .init.data section
as it is now only needed at boot up.

Signed-off-by: Rafael Aquini <aquini@redhat.com>