spec: latest bfa driver requires newer firmware
Patchwork-id: 65129
O-Subject: [RHEL6 PATCH] redhat: spec: latest bfa driver requires newer firmware
Bugzilla:
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: Jeff Burke <jburke@redhat.com>

This is a necessary follow-up patch for
https://bugzilla.redhat.com/show_bug.cgi?id=928983

The bfa driver update in kernel-2.6.32-408.el6 fails to install on systems
with the old bfa-firmware. To ensure people don't install and reboot with
the older firmware in place, we need to update the Conflicts marker
to the kernel rpm, to trigger bfa-firmware also upgrading when a newer
kernel that needs it is installed.

Signed-off-by: Rafael Aquini <aquini@redhat.com>