ixgbe: fix broken module parameter FdirMode
Patchwork-id: 65400
O-Subject: [PATCH RHEL6.5 2/2] ixgbe: fix broken module parameter FdirMode
Bugzilla: 958967
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

Significant changes to Flow Director initialization in the ixgbe driver
during RHEL6.4 development broke the ability to disable the Flow
Director when requested.

This patch keeps the Flow Director disabled when requested and cleans up
the output of modinfo to make it clear that this completely enables or
disables the Flow Director rather than implying that it only controls
hashing mode and not the perfect filter.

This was verified by bmarson and me and will resolve RHBZ 958967.

Signed-off-by: Rafael Aquini <aquini@redhat.com>